refactor(webdav): extract mutation operations into dedicated submodule
Extract WebDAV mutation methods from backend/mod.rs into a new mutation/
submodule for improved organization and maintainability.

Changes:
- Create mutation/ submodule with focused modules:
  - audit.rs: audit logging helpers for mutations
  - delete.rs: delete_with_locks operation
  - move_node.rs: move_with_locks operation
  - copy.rs: copy_file_with_locks operation
  - collection.rs: prepare_collection_with_locks operation
- Move 934 lines of mutation logic from mod.rs to specialized files
- Preserve all behavior, error handling, and transaction boundaries
- Maintain existing public API surface unchanged
- Keep helper visibility (pub(super)) for audit logging methods

No functional changes, pure code reorganization.

Refs: #466
