+++ src/webdav/backend/mod.rs
@@ -5,12 +5,13 @@ mod download_audit;
 pub mod file;
 pub mod lock;
 mod metadata;
+mod mutation;
 pub mod path_resolver;
 
 use aster_forge_db::transaction;
 use std::{collections::HashMap, pin::Pin};
 
-use sea_orm::{ActiveModelTrait, ConnectionTrait, Set};
+use sea_orm::ConnectionTrait;
 use tokio::io::{AsyncRead, AsyncReadExt};
 
 use crate::db::repository::{file_repo, folder_repo, property_repo, team_repo, user_repo};
@@ -40,9 +41,8 @@ use aster_forge_webdav::{
     DavBackendError, DavBackendErrorKind, DavConditionalOutcome, DavConditionalResource,
     DavContentStream, DavDirectoryEnumerator, DavDirectoryPage, DavDirectoryPageRequest,
     DavDownloadSource, DavFileSystem, DavLockError, DavMetaData, DavMutationCredentials,
-    DavMutationOperation, DavMutationTargetRole, DavOpenedDownload, DavPath, DavProp,
-    DavWriteOptions, DavWriteSystem, FsError, FsFuture, IfHeader, parent_relative_path,
-    plan_http_conditionals,
+    DavOpenedDownload, DavPath, DavProp, DavWriteOptions, DavWriteSystem, FsError, FsFuture,
+    IfHeader, plan_http_conditionals,
 };
 
 /// AsterDrive WebDAV 文件系统，per-account workspace 实例。
@@ -209,940 +209,6 @@ impl AsterDavFs {
                 Ok(AsterDavMeta::from_file(&file, &blob))
             }
         }
-    }
-
-    /// Deletes one resource and all locks rooted in its namespace in one writer commit.
-    pub(crate) async fn delete_with_locks(
-        &self,
-        path: &DavPath,
-        is_collection: bool,
-        operation: DavMutationOperation,
-        role: DavMutationTargetRole,
-        conditions: DavMutationConditions<'_>,
-    ) -> Result<(), AsterDavMutationError> {
-        let txn = transaction::begin(self.state.writer_db())
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        let mut lock_mutation = lock::WebDavLockMutation::begin(&txn, self.scope)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-        lock_mutation
-            .lock_ancestor_entities(self.scope, self.root_folder_id, path)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-
-        let node =
-            path_resolver::resolve_path_in_scope(&txn, self.scope, path, self.root_folder_id)
-                .await
-                .map_err(AsterDavMutationError::FileSystem)?;
-
-        enum Deleted {
-            File(file_entity::Model),
-            Folder(folder::FolderTreeDeletion),
-        }
-
-        let prepared_folder = match (&node, is_collection) {
-            (ResolvedNode::Folder(folder), true) => Some(
-                folder::lock_tree_for_deletion_on(
-                    &txn,
-                    self.scope,
-                    folder.id,
-                    Some(MUTATION_FOLDER_TREE_LIMITS),
-                    true,
-                )
-                .await
-                .map_err(to_fs_error)
-                .map_err(AsterDavMutationError::FileSystem)?,
-            ),
-            (ResolvedNode::File(_), false) => None,
-            _ => return Err(AsterDavMutationError::FileSystem(FsError::Forbidden)),
-        };
-
-        revalidate_atomic_target(
-            &txn,
-            lock_mutation.namespace_id(),
-            self.scope,
-            self.root_folder_id,
-            AtomicTargetRevalidation {
-                path,
-                check_locks: true,
-                deep: is_collection,
-            },
-            &conditions,
-        )
-        .await?;
-        if let Some(parent) = parent_relative_path(path.as_str())
-            && let Ok(parent) = DavPath::new(&parent)
-        {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: &parent,
-                    check_locks: true,
-                    deep: false,
-                },
-                &conditions,
-            )
-            .await?;
-        }
-
-        let deleted = match (node, prepared_folder) {
-            (ResolvedNode::File(file), None) => {
-                file_ops::delete_in_scope_on(&txn, self.scope, file.id, true)
-                    .await
-                    .map_err(to_fs_error)
-                    .map_err(AsterDavMutationError::FileSystem)?;
-                Deleted::File(file)
-            }
-            (ResolvedNode::Folder(_), Some(locked)) => Deleted::Folder(
-                folder::apply_locked_tree_deletion_on(&txn, locked, chrono::Utc::now())
-                    .await
-                    .map_err(to_fs_error)
-                    .map_err(AsterDavMutationError::FileSystem)?,
-            ),
-            _ => return Err(AsterDavMutationError::FileSystem(FsError::Forbidden)),
-        };
-        lock_mutation
-            .delete_rooted_locks(path)
-            .await
-            .map_err(map_atomic_lock_error)?;
-        lock_mutation
-            .finish()
-            .await
-            .map_err(map_atomic_lock_error)?;
-        transaction::commit(txn)
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-
-        match deleted {
-            Deleted::File(file) => {
-                storage_change::publish(
-                    &self.state,
-                    storage_change::StorageChangeEvent::new(
-                        storage_change::StorageChangeKind::FileTrashed,
-                        self.scope,
-                        vec![file.id],
-                        vec![],
-                        vec![file.folder_id],
-                    ),
-                );
-                if operation == DavMutationOperation::Delete
-                    || role == DavMutationTargetRole::Destination
-                {
-                    self.log_deleted_file(path, &file).await;
-                }
-            }
-            Deleted::Folder(outcome) => {
-                storage_change::publish(
-                    &self.state,
-                    storage_change::StorageChangeEvent::new(
-                        storage_change::StorageChangeKind::FolderTrashed,
-                        self.scope,
-                        vec![],
-                        vec![outcome.folder.id],
-                        vec![outcome.folder.parent_id],
-                    ),
-                );
-                folder::invalidate_folder_path_cache_for_ids(&self.state, &[outcome.folder.id])
-                    .await;
-                if operation == DavMutationOperation::Delete
-                    || role == DavMutationTargetRole::Destination
-                {
-                    self.log_deleted_folder(path, &outcome.folder).await;
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Moves one resource, destination overwrite state, rooted locks, and derived lock flags in
-    /// one writer commit.
-    pub(crate) async fn move_with_locks(
-        &self,
-        source: &DavPath,
-        destination: &DavPath,
-        conditions: DavMutationConditions<'_>,
-    ) -> Result<(), AsterDavMutationError> {
-        let txn = transaction::begin(self.state.writer_db())
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        let mut lock_mutation = lock::WebDavLockMutation::begin(&txn, self.scope)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-        let mut lock_paths = [source, destination];
-        lock_paths.sort_by(|left, right| left.as_str().cmp(right.as_str()));
-        for path in lock_paths {
-            lock_mutation
-                .lock_ancestor_entities(self.scope, self.root_folder_id, path)
-                .await
-                .map_err(map_ancestor_lock_error)?;
-        }
-
-        let source_node =
-            path_resolver::resolve_path_in_scope(&txn, self.scope, source, self.root_folder_id)
-                .await
-                .map_err(AsterDavMutationError::FileSystem)?;
-        let destination_node = match path_resolver::resolve_path_in_scope(
-            &txn,
-            self.scope,
-            destination,
-            self.root_folder_id,
-        )
-        .await
-        {
-            Ok(node) => Some(node),
-            Err(FsError::NotFound) => None,
-            Err(error) => return Err(AsterDavMutationError::FileSystem(error)),
-        };
-        let overwritten = destination_node.as_ref().and_then(|node| match node {
-            ResolvedNode::File(file) => Some(DeletedResource::File(file.clone())),
-            ResolvedNode::Folder(folder) => Some(DeletedResource::Folder(folder.clone())),
-            ResolvedNode::Root => None,
-        });
-
-        let source_is_collection = matches!(source_node, ResolvedNode::Folder(_));
-        let _source_tree_lock = match &source_node {
-            ResolvedNode::Folder(folder) => Some(
-                folder::lock_tree_for_deletion_on(
-                    &txn,
-                    self.scope,
-                    folder.id,
-                    Some(MUTATION_FOLDER_TREE_LIMITS),
-                    true,
-                )
-                .await
-                .map_err(to_fs_error)
-                .map_err(AsterDavMutationError::FileSystem)?,
-            ),
-            ResolvedNode::File(_) => None,
-            ResolvedNode::Root => {
-                return Err(AsterDavMutationError::FileSystem(FsError::Forbidden));
-            }
-        };
-        revalidate_atomic_target(
-            &txn,
-            lock_mutation.namespace_id(),
-            self.scope,
-            self.root_folder_id,
-            AtomicTargetRevalidation {
-                path: source,
-                check_locks: true,
-                deep: source_is_collection,
-            },
-            &conditions,
-        )
-        .await?;
-        if let Some(parent) = parent_relative_path(source.as_str())
-            && let Ok(parent) = DavPath::new(&parent)
-        {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: &parent,
-                    check_locks: true,
-                    deep: false,
-                },
-                &conditions,
-            )
-            .await?;
-        }
-
-        let prepared_destination_folder = match &destination_node {
-            Some(ResolvedNode::Folder(folder)) => Some(
-                folder::lock_tree_for_deletion_on(
-                    &txn,
-                    self.scope,
-                    folder.id,
-                    Some(MUTATION_FOLDER_TREE_LIMITS),
-                    true,
-                )
-                .await
-                .map_err(to_fs_error)
-                .map_err(AsterDavMutationError::FileSystem)?,
-            ),
-            _ => None,
-        };
-        if let Some(node) = &destination_node {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: destination,
-                    check_locks: true,
-                    deep: matches!(node, ResolvedNode::Folder(_)),
-                },
-                &conditions,
-            )
-            .await?;
-        }
-        if let Some(parent) = parent_relative_path(destination.as_str())
-            && let Ok(parent) = DavPath::new(&parent)
-        {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: &parent,
-                    check_locks: true,
-                    deep: false,
-                },
-                &conditions,
-            )
-            .await?;
-        }
-
-        if let Some(node) = destination_node {
-            match (node, prepared_destination_folder) {
-                (ResolvedNode::File(file), None) => {
-                    file_ops::delete_in_scope_on(&txn, self.scope, file.id, true)
-                        .await
-                        .map_err(to_fs_error)
-                        .map_err(AsterDavMutationError::FileSystem)?;
-                }
-                (ResolvedNode::Folder(_), Some(locked)) => {
-                    folder::apply_locked_tree_deletion_on(&txn, locked, chrono::Utc::now())
-                        .await
-                        .map_err(to_fs_error)
-                        .map_err(AsterDavMutationError::FileSystem)?;
-                }
-                _ => return Err(AsterDavMutationError::FileSystem(FsError::Forbidden)),
-            }
-            lock_mutation
-                .delete_descendant_rooted_locks(destination)
-                .await
-                .map_err(map_atomic_lock_error)?;
-        }
-
-        let (destination_parent_id, destination_name) = path_resolver::resolve_parent_in_scope(
-            &txn,
-            self.scope,
-            destination,
-            self.root_folder_id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?;
-        enum Moved {
-            File {
-                previous: file_entity::Model,
-                current: file_entity::Model,
-            },
-            Folder {
-                previous: aster_drive_model::entities::folder::Model,
-                current: aster_drive_model::entities::folder::Model,
-            },
-        }
-
-        let (entity_type, entity_id, event, moved) = match source_node {
-            ResolvedNode::File(file) => {
-                let mut active: file_entity::ActiveModel = file.clone().into();
-                active.name = Set(destination_name);
-                active.folder_id = Set(destination_parent_id);
-                active.updated_at = Set(chrono::Utc::now());
-                let moved = active
-                    .update(&txn)
-                    .await
-                    .map_err(crate::errors::AsterError::from)
-                    .map_err(to_fs_error)
-                    .map_err(AsterDavMutationError::FileSystem)?;
-                (
-                    EntityType::File,
-                    moved.id,
-                    storage_change::StorageChangeEvent::new(
-                        storage_change::StorageChangeKind::FileUpdated,
-                        self.scope,
-                        vec![moved.id],
-                        vec![],
-                        vec![file.folder_id, moved.folder_id],
-                    ),
-                    Moved::File {
-                        previous: file,
-                        current: moved,
-                    },
-                )
-            }
-            ResolvedNode::Folder(folder) => {
-                let mut active: aster_drive_model::entities::folder::ActiveModel =
-                    folder.clone().into();
-                active.name = Set(destination_name);
-                active.parent_id = Set(destination_parent_id);
-                active.updated_at = Set(chrono::Utc::now());
-                let moved = active
-                    .update(&txn)
-                    .await
-                    .map_err(crate::errors::AsterError::from)
-                    .map_err(to_fs_error)
-                    .map_err(AsterDavMutationError::FileSystem)?;
-                (
-                    EntityType::Folder,
-                    moved.id,
-                    storage_change::StorageChangeEvent::new(
-                        storage_change::StorageChangeKind::FolderUpdated,
-                        self.scope,
-                        vec![],
-                        vec![moved.id],
-                        vec![folder.parent_id, moved.parent_id],
-                    ),
-                    Moved::Folder {
-                        previous: folder,
-                        current: moved,
-                    },
-                )
-            }
-            ResolvedNode::Root => {
-                return Err(AsterDavMutationError::FileSystem(FsError::Forbidden));
-            }
-        };
-        lock_mutation
-            .delete_rooted_locks(source)
-            .await
-            .map_err(map_atomic_lock_error)?;
-        lock_mutation
-            .rebind_destination_root_locks(destination, entity_type, entity_id)
-            .await
-            .map_err(map_atomic_lock_error)?;
-        lock_mutation
-            .finish()
-            .await
-            .map_err(map_atomic_lock_error)?;
-        transaction::commit(txn)
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        storage_change::publish(&self.state, event);
-        if let Some(overwritten) = overwritten.as_ref() {
-            self.log_deleted_resource(destination, overwritten).await;
-        }
-        match moved {
-            Moved::File { previous, current } => {
-                self.log_file_transfer(
-                    audit::AuditAction::FileMove,
-                    source,
-                    destination,
-                    &previous,
-                    &current,
-                )
-                .await;
-            }
-            Moved::Folder { previous, current } => {
-                self.log_folder_transfer(
-                    audit::AuditAction::FolderMove,
-                    source,
-                    destination,
-                    &previous,
-                    &current,
-                )
-                .await;
-            }
-        }
-        if entity_type == EntityType::Folder {
-            folder::invalidate_folder_path_cache_for_ids(&self.state, &[entity_id]).await;
-        }
-        Ok(())
-    }
-
-    /// Copies one file and atomically preserves/rebinds locks rooted at an overwritten destination.
-    pub(crate) async fn copy_file_with_locks(
-        &self,
-        source: &DavPath,
-        destination: &DavPath,
-        conditions: DavMutationConditions<'_>,
-    ) -> Result<(), AsterDavMutationError> {
-        let txn = transaction::begin(self.state.writer_db())
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        let mut lock_mutation = lock::WebDavLockMutation::begin(&txn, self.scope)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-        lock_mutation
-            .lock_ancestor_entities(self.scope, self.root_folder_id, destination)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-        let source_file = match path_resolver::resolve_path_in_scope(
-            &txn,
-            self.scope,
-            source,
-            self.root_folder_id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?
-        {
-            ResolvedNode::File(file) => file_repo::lock_by_id(&txn, file.id)
-                .await
-                .map_err(to_fs_error)
-                .map_err(AsterDavMutationError::FileSystem)?,
-            _ => return Err(AsterDavMutationError::FileSystem(FsError::Forbidden)),
-        };
-        revalidate_atomic_target(
-            &txn,
-            lock_mutation.namespace_id(),
-            self.scope,
-            self.root_folder_id,
-            AtomicTargetRevalidation {
-                path: source,
-                check_locks: false,
-                deep: false,
-            },
-            &conditions,
-        )
-        .await?;
-        let destination_node = match path_resolver::resolve_path_in_scope(
-            &txn,
-            self.scope,
-            destination,
-            self.root_folder_id,
-        )
-        .await
-        {
-            Ok(node) => Some(node),
-            Err(FsError::NotFound) => None,
-            Err(error) => return Err(AsterDavMutationError::FileSystem(error)),
-        };
-        let overwritten = destination_node.as_ref().and_then(|node| match node {
-            ResolvedNode::File(file) => Some(DeletedResource::File(file.clone())),
-            ResolvedNode::Folder(folder) => Some(DeletedResource::Folder(folder.clone())),
-            ResolvedNode::Root => None,
-        });
-        let prepared_destination_folder = match &destination_node {
-            Some(ResolvedNode::Folder(folder)) => Some(
-                folder::lock_tree_for_deletion_on(
-                    &txn,
-                    self.scope,
-                    folder.id,
-                    Some(MUTATION_FOLDER_TREE_LIMITS),
-                    true,
-                )
-                .await
-                .map_err(to_fs_error)
-                .map_err(AsterDavMutationError::FileSystem)?,
-            ),
-            _ => None,
-        };
-        if let Some(node) = &destination_node {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: destination,
-                    check_locks: true,
-                    deep: matches!(node, ResolvedNode::Folder(_)),
-                },
-                &conditions,
-            )
-            .await?;
-        }
-        if let Some(parent) = parent_relative_path(destination.as_str())
-            && let Ok(parent) = DavPath::new(&parent)
-        {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: &parent,
-                    check_locks: true,
-                    deep: false,
-                },
-                &conditions,
-            )
-            .await?;
-        }
-        if let Some(node) = destination_node {
-            match (node, prepared_destination_folder) {
-                (ResolvedNode::File(file), None) => {
-                    file_ops::delete_in_scope_on(&txn, self.scope, file.id, true)
-                        .await
-                        .map_err(to_fs_error)
-                        .map_err(AsterDavMutationError::FileSystem)?;
-                }
-                (ResolvedNode::Folder(_), Some(locked)) => {
-                    folder::apply_locked_tree_deletion_on(&txn, locked, chrono::Utc::now())
-                        .await
-                        .map_err(to_fs_error)
-                        .map_err(AsterDavMutationError::FileSystem)?;
-                }
-                _ => return Err(AsterDavMutationError::FileSystem(FsError::Forbidden)),
-            }
-            lock_mutation
-                .delete_descendant_rooted_locks(destination)
-                .await
-                .map_err(map_atomic_lock_error)?;
-        }
-        let (destination_parent_id, destination_name) = path_resolver::resolve_parent_in_scope(
-            &txn,
-            self.scope,
-            destination,
-            self.root_folder_id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?;
-        let copied = file_ops::duplicate_file_record_in_scope_on(
-            &txn,
-            self.scope,
-            &source_file,
-            destination_parent_id,
-            &destination_name,
-        )
-        .await
-        .map_err(to_fs_error)
-        .map_err(AsterDavMutationError::FileSystem)?;
-        copy_visible_entity_properties_on(
-            &txn,
-            EntityType::File,
-            source_file.id,
-            EntityType::File,
-            copied.id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?;
-        lock_mutation
-            .rebind_destination_root_locks(destination, EntityType::File, copied.id)
-            .await
-            .map_err(map_atomic_lock_error)?;
-        lock_mutation
-            .finish()
-            .await
-            .map_err(map_atomic_lock_error)?;
-        transaction::commit(txn)
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        storage_change::publish(
-            &self.state,
-            storage_change::StorageChangeEvent::new(
-                storage_change::StorageChangeKind::FileCreated,
-                self.scope,
-                vec![copied.id],
-                vec![],
-                vec![copied.folder_id],
-            )
-            .with_storage_delta(copied.size),
-        );
-        if let Some(overwritten) = overwritten.as_ref() {
-            self.log_deleted_resource(destination, overwritten).await;
-        }
-        self.log_file_transfer(
-            audit::AuditAction::FileCopy,
-            source,
-            destination,
-            &source_file,
-            &copied,
-        )
-        .await;
-        Ok(())
-    }
-
-    /// Prepares one collection destination for recursive COPY/MOVE while preserving the existing
-    /// destination lock scope in the same writer commit.
-    pub(crate) async fn prepare_collection_with_locks(
-        &self,
-        source: &DavPath,
-        destination: &DavPath,
-        operation: DavMutationOperation,
-        conditions: DavMutationConditions<'_>,
-    ) -> Result<(), AsterDavMutationError> {
-        let txn = transaction::begin(self.state.writer_db())
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        let mut lock_mutation = lock::WebDavLockMutation::begin(&txn, self.scope)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-        lock_mutation
-            .lock_ancestor_entities(self.scope, self.root_folder_id, destination)
-            .await
-            .map_err(map_ancestor_lock_error)?;
-        let source_folder = match path_resolver::resolve_path_in_scope(
-            &txn,
-            self.scope,
-            source,
-            self.root_folder_id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?
-        {
-            ResolvedNode::Folder(folder) => folder_repo::lock_by_id(&txn, folder.id)
-                .await
-                .map_err(to_fs_error)
-                .map_err(AsterDavMutationError::FileSystem)?,
-            _ => return Err(AsterDavMutationError::FileSystem(FsError::Forbidden)),
-        };
-        revalidate_atomic_target(
-            &txn,
-            lock_mutation.namespace_id(),
-            self.scope,
-            self.root_folder_id,
-            AtomicTargetRevalidation {
-                path: source,
-                check_locks: operation == DavMutationOperation::Move,
-                deep: false,
-            },
-            &conditions,
-        )
-        .await?;
-        let destination_node = match path_resolver::resolve_path_in_scope(
-            &txn,
-            self.scope,
-            destination,
-            self.root_folder_id,
-        )
-        .await
-        {
-            Ok(node) => Some(node),
-            Err(FsError::NotFound) => None,
-            Err(error) => return Err(AsterDavMutationError::FileSystem(error)),
-        };
-        let overwritten = destination_node.as_ref().and_then(|node| match node {
-            ResolvedNode::File(file) => Some(DeletedResource::File(file.clone())),
-            ResolvedNode::Folder(_) | ResolvedNode::Root => None,
-        });
-        if destination_node.is_some() {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: destination,
-                    check_locks: true,
-                    deep: false,
-                },
-                &conditions,
-            )
-            .await?;
-        }
-        if let Some(parent) = parent_relative_path(destination.as_str())
-            && let Ok(parent) = DavPath::new(&parent)
-        {
-            revalidate_atomic_target(
-                &txn,
-                lock_mutation.namespace_id(),
-                self.scope,
-                self.root_folder_id,
-                AtomicTargetRevalidation {
-                    path: &parent,
-                    check_locks: true,
-                    deep: false,
-                },
-                &conditions,
-            )
-            .await?;
-        }
-        let (destination_parent_id, destination_name) = path_resolver::resolve_parent_in_scope(
-            &txn,
-            self.scope,
-            destination,
-            self.root_folder_id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?;
-        let destination_folder = match destination_node {
-            Some(ResolvedNode::Folder(folder)) => Some(folder),
-            Some(ResolvedNode::File(file)) => {
-                file_ops::delete_in_scope_on(&txn, self.scope, file.id, true)
-                    .await
-                    .map_err(to_fs_error)
-                    .map_err(AsterDavMutationError::FileSystem)?;
-                None
-            }
-            None => None,
-            Some(ResolvedNode::Root) => {
-                return Err(AsterDavMutationError::FileSystem(FsError::Forbidden));
-            }
-        };
-        let (destination_folder, created) = if let Some(folder) = destination_folder {
-            (folder, false)
-        } else {
-            let created_by_username =
-                crate::services::workspace::storage::load_scope_actor_username(&txn, self.scope)
-                    .await
-                    .map_err(to_fs_error)
-                    .map_err(AsterDavMutationError::FileSystem)?;
-            let now = chrono::Utc::now();
-            let folder = folder_repo::create(
-                &txn,
-                aster_drive_model::entities::folder::ActiveModel {
-                    name: Set(destination_name),
-                    parent_id: Set(destination_parent_id),
-                    team_id: Set(self.scope.team_id()),
-                    owner_user_id: Set(self.scope.owner_user_id()),
-                    created_by_user_id: Set(Some(self.scope.actor_user_id())),
-                    created_by_username: Set(created_by_username),
-                    policy_id: Set(source_folder.policy_id),
-                    created_at: Set(now),
-                    updated_at: Set(now),
-                    ..Default::default()
-                },
-            )
-            .await
-            .map_err(to_fs_error)
-            .map_err(AsterDavMutationError::FileSystem)?;
-            (folder, true)
-        };
-        copy_visible_entity_properties_on(
-            &txn,
-            EntityType::Folder,
-            source_folder.id,
-            EntityType::Folder,
-            destination_folder.id,
-        )
-        .await
-        .map_err(AsterDavMutationError::FileSystem)?;
-        lock_mutation
-            .rebind_destination_root_locks(destination, EntityType::Folder, destination_folder.id)
-            .await
-            .map_err(map_atomic_lock_error)?;
-        lock_mutation
-            .finish()
-            .await
-            .map_err(map_atomic_lock_error)?;
-        transaction::commit(txn)
-            .await
-            .map_err(|_| AsterDavMutationError::Backend)?;
-        storage_change::publish(
-            &self.state,
-            storage_change::StorageChangeEvent::new(
-                if created {
-                    storage_change::StorageChangeKind::FolderCreated
-                } else {
-                    storage_change::StorageChangeKind::FolderUpdated
-                },
-                self.scope,
-                vec![],
-                vec![destination_folder.id],
-                vec![destination_folder.parent_id],
-            ),
-        );
-        if let Some(overwritten) = overwritten.as_ref() {
-            self.log_deleted_resource(destination, overwritten).await;
-        }
-        self.log_folder_transfer(
-            if operation == DavMutationOperation::Move {
-                audit::AuditAction::FolderMove
-            } else {
-                audit::AuditAction::FolderCopy
-            },
-            source,
-            destination,
-            &source_folder,
-            &destination_folder,
-        )
-        .await;
-        Ok(())
-    }
-
-    async fn log_deleted_resource(&self, path: &DavPath, resource: &DeletedResource) {
-        match resource {
-            DeletedResource::File(file) => self.log_deleted_file(path, file).await,
-            DeletedResource::Folder(folder) => self.log_deleted_folder(path, folder).await,
-        }
-    }
-
-    async fn log_deleted_file(&self, path: &DavPath, file: &file_entity::Model) {
-        let details = Some(serde_json::json!({
-            "folder_id": file.folder_id,
-            "path": path.as_str(),
-            "team_id": self.scope.team_id(),
-        }));
-        audit::log_with_details(
-            &self.state,
-            &self.audit_ctx,
-            audit::AuditAction::FileDelete,
-            crate::services::ops::audit::AuditEntityType::File,
-            Some(file.id),
-            Some(&file.name),
-            || details.clone(),
-        )
-        .await;
-    }
-
-    async fn log_deleted_folder(
-        &self,
-        path: &DavPath,
-        folder: &aster_drive_model::entities::folder::Model,
-    ) {
-        let details = Some(serde_json::json!({
-            "parent_id": folder.parent_id,
-            "path": path.as_str(),
-            "team_id": self.scope.team_id(),
-        }));
-        audit::log_with_details(
-            &self.state,
-            &self.audit_ctx,
-            audit::AuditAction::FolderDelete,
-            crate::services::ops::audit::AuditEntityType::Folder,
-            Some(folder.id),
-            Some(&folder.name),
-            || details.clone(),
-        )
-        .await;
-    }
-
-    async fn log_file_transfer(
-        &self,
-        action: audit::AuditAction,
-        source: &DavPath,
-        destination: &DavPath,
-        previous: &file_entity::Model,
-        current: &file_entity::Model,
-    ) {
-        let details = Some(serde_json::json!({
-            "source_folder_id": previous.folder_id,
-            "source_path": source.as_str(),
-            "target_folder_id": current.folder_id,
-            "target_path": destination.as_str(),
-            "previous_name": previous.name,
-            "next_name": current.name,
-            "team_id": self.scope.team_id(),
-        }));
-        audit::log_with_details(
-            &self.state,
-            &self.audit_ctx,
-            action,
-            crate::services::ops::audit::AuditEntityType::File,
-            Some(current.id),
-            Some(&current.name),
-            || details.clone(),
-        )
-        .await;
-    }
-
-    async fn log_folder_transfer(
-        &self,
-        action: audit::AuditAction,
-        source: &DavPath,
-        destination: &DavPath,
-        previous: &aster_drive_model::entities::folder::Model,
-        current: &aster_drive_model::entities::folder::Model,
-    ) {
-        let details = Some(serde_json::json!({
-            "source_parent_id": previous.parent_id,
-            "source_path": source.as_str(),
-            "target_parent_id": current.parent_id,
-            "target_path": destination.as_str(),
-            "previous_name": previous.name,
-            "next_name": current.name,
-            "team_id": self.scope.team_id(),
-        }));
-        audit::log_with_details(
-            &self.state,
-            &self.audit_ctx,
-            action,
-            crate::services::ops::audit::AuditEntityType::Folder,
-            Some(current.id),
-            Some(&current.name),
-            || details.clone(),
-        )
-        .await;
     }
 
     pub(crate) async fn open_download_stream_for_file(
